requirements.txt
icalendar
python-dateutil
selenium
beautifulsoup4
webdriver-manager
requests
pyyaml
